audio_pipeline/requirements.txt
torch
torchaudio
speechbrain
nemo_toolkit[asr]
chromadb
fastapi
uvicorn
python-multipart
pydub
librosa
